refactor Scheduler and dispatcher
